abstract permission class updated, all test updated as per new changes, added user to save method

+++ src/resource_manager/services.py
@@ -7,7 +7,11 @@ from resource_calendar.models import WeeklyShiftTemplate
 from resource_manager.models import Resource, ResourcePool, WorkUnit
 
 
-class ResourceService(AbstractPermissionService):
+class ResourceService:
+    def __init__(self, user) -> None:
+        self.user = user
+        self.permission_service = AbstractPermissionService(user=user)
+
     def create(
         self,
         *,
@@ -21,7 +25,7 @@ class ResourceService(AbstractPermissionService):
         weekly_shift_template: WeeklyShiftTemplate = None,
     ) -> Resource:
         # check permissions for add resource
-        if not self.check_for_permission("add_resource"):
+        if not self.permission_service.check_for_permission("add_resource"):
             raise PermissionDenied()
 
         resource = Resource.objects.create(
@@ -48,7 +52,7 @@ class ResourceService(AbstractPermissionService):
 
     def update(self, *, instance: Resource, data: dict) -> Resource:
         # check permissions for update resource
-        if not self.check_for_permission("change_resource"):
+        if not self.permission_service.check_for_permission("change_resource"):
             raise PermissionDenied()
 
         fields = [
@@ -69,15 +73,16 @@ class ResourceService(AbstractPermissionService):
 
     def delete(self, instance: Resource) -> None:
         # check permissions for delete resource
-        if not self.check_for_permission("delete_resource"):
+        if not self.permission_service.check_for_permission("delete_resource"):
             raise PermissionDenied()
 
         instance.delete()
 
 
 class WorkUnitService:
-    def __init__(self):
-        pass
+    def __init__(self, user) -> None:
+        self.user = user
+        self.permission_service = AbstractPermissionService(user=user)
 
     def create(
         self,
@@ -88,6 +93,10 @@ class WorkUnitService:
         resources: list[Resource] = None,
         resource_pools: list[ResourcePool] = None,
     ) -> WorkUnit:
+        # check permissions for add workunit
+        if not self.permission_service.check_for_permission("add_workunit"):
+            raise PermissionDenied()
+
         work_unit = WorkUnit.objects.create(
             name=name,
             external_id=external_id,
@@ -95,7 +104,7 @@ class WorkUnitService:
         )
 
         work_unit.full_clean()
-        work_unit.save()
+        work_unit.save(user=self.user)
 
         if resources:
             work_unit.resources.set(resources)
@@ -106,6 +115,10 @@ class WorkUnitService:
         return work_unit
 
     def update(self, *, instance: WorkUnit, data: dict) -> WorkUnit:
+        # check permissions for update work unit
+        if not self.permission_service.check_for_permission("change_workunit"):
+            raise PermissionDenied()
+
         fields = [
             "name",
             "external_id",
@@ -114,17 +127,24 @@ class WorkUnitService:
             "resource_pools",
         ]
 
-        work_unit, _ = model_update(instance=instance, fields=fields, data=data)
+        work_unit, _ = model_update(
+            instance=instance, fields=fields, data=data, user=self.user
+        )
 
         return work_unit
 
     def delete(self, instance: WorkUnit) -> None:
+        # check permission for delete work unit
+        if not self.permission_service.check_for_permission("delete_workunit"):
+            raise PermissionDenied()
+
         instance.delete()
 
 
 class ResourcePoolService:
-    def __init__(self):
-        pass
+    def __init__(self, user) -> None:
+        self.user = user
+        self.permission_service = AbstractPermissionService(user=user)
 
     def create(
         self,
@@ -135,6 +155,10 @@ class ResourcePoolService:
         resources: list[Resource] = None,
         work_units: list[WorkUnit] = None,
     ) -> ResourcePool:
+        # check permissions for add resource pool
+        if not self.permission_service.check_for_permission("add_resourcepool"):
+            raise PermissionDenied()
+
         resource_pool = ResourcePool.objects.create(
             name=name,
             external_id=external_id,
@@ -142,7 +166,7 @@ class ResourcePoolService:
         )
 
         resource_pool.full_clean()
-        resource_pool.save()
+        resource_pool.save(user=self.user)
 
         if resources:
             resource_pool.resources.set(resources)
@@ -153,6 +177,10 @@ class ResourcePoolService:
         return resource_pool
 
     def update(self, *, instance: ResourcePool, data: dict) -> ResourcePool:
+        # check permissions for update resource pool
+        if not self.permission_service.check_for_permission("change_resourcepool"):
+            raise PermissionDenied()
+
         fields = [
             "name",
             "external_id",
@@ -166,4 +194,8 @@ class ResourcePoolService:
         return resource_pool
 
     def delete(self, instance: ResourcePool) -> None:
+        # check permissions for delete resource pool
+        if not self.permission_service.check_for_permission("delete_resourcepool"):
+            raise PermissionDenied()
+
         instance.delete()
